SRP: WeatherForecastModel, WeatherDetailView 분리
