+++ WeatherForecast/WeatherForecast/ViewController.swift
@@ -8,21 +8,13 @@ import UIKit
 
 class ViewController: UIViewController {
     var tableView: UITableView!
-    let refreshControl: UIRefreshControl = UIRefreshControl()
-    var weatherJSON: WeatherJSON?
-    var icons: [UIImage]?
     let imageCache: ImageCache = ImageCache()
-    let dateFormatter: DateFormatter = {
-        let formatter: DateFormatter = DateFormatter()
-        formatter.locale = .init(identifier: "ko_KR")
-        formatter.dateFormat = "yyyy-MM-dd(EEEEE) a HH:mm"
-        return formatter
-    }()
-    
+    var model: WeatherForecastModel!
     var tempUnit: TempUnit = .metric
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        model = WeatherForecastModel()
         initialSetUp()
     }
 }
@@ -41,9 +33,9 @@ extension ViewController {
     }
     
     @objc private func refresh() {
-        fetchWeatherJSON()
+        requestWeatherJSON()
         tableView.reloadData()
-        refreshControl.endRefreshing()
+        tableView.refreshControl?.endRefreshing()
     }
     
     private func initialSetUp() {
@@ -51,11 +43,10 @@ extension ViewController {
         
         layTable()
         
-        refreshControl.addTarget(self,
-                                 action: #selector(refresh),
-                                 for: .valueChanged)
-        
-        tableView.refreshControl = refreshControl
+        tableView.refreshControl = UIRefreshControl()
+        tableView.refreshControl?.addTarget(self,
+                                           action: #selector(refresh),
+                                           for: .valueChanged)
         tableView.register(WeatherTableViewCell.self, forCellReuseIdentifier: "WeatherCell")
         tableView.dataSource = self
         tableView.delegate = self
@@ -78,51 +69,32 @@ extension ViewController {
 }
 
 extension ViewController {
-    private func fetchWeatherJSON() {
-        
-        let jsonDecoder: JSONDecoder = .init()
-        jsonDecoder.keyDecodingStrategy = .convertFromSnakeCase
-
-        guard let data = NSDataAsset(name: "weather")?.data else {
-            return
-        }
-        
-        let info: WeatherJSON
-        do {
-            info = try jsonDecoder.decode(WeatherJSON.self, from: data)
-        } catch {
-            print(error.localizedDescription)
-            return
-        }
-
-        weatherJSON = info
-        navigationItem.title = weatherJSON?.city.name
+    private func requestWeatherJSON() {
+        navigationItem.title = model.weatherJSON.city.name
     }
 }
 
 extension ViewController: UITableViewDataSource {
     
-    func numberOfSections(in tableView: UITableView) -> Int {
-        1
-    }
-    
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        weatherJSON?.weatherForecast.count ?? 0
+        model.weatherJSON?.weatherForecast.count ?? 0
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell: UITableViewCell = tableView.dequeueReusableCell(withIdentifier: "WeatherCell", for: indexPath)
         
         guard let cell: WeatherTableViewCell = cell as? WeatherTableViewCell,
-              let weatherForecastInfo = weatherJSON?.weatherForecast[indexPath.row] else {
+              let weatherForecastInfo = model.weatherJSON?.weatherForecast[indexPath.row] else {
             return cell
         }
         
-        cell.weatherLabel.text = weatherForecastInfo.weather.main
-        cell.descriptionLabel.text = weatherForecastInfo.weather.description
-        cell.temperatureLabel.text = "\(weatherForecastInfo.main.temp)\(tempUnit.expression)"
+        cell.setData(weatherForecastInfo, unit: tempUnit)
         
         let date: Date = Date(timeIntervalSince1970: weatherForecastInfo.dt)
+        let dateFormatter: DateFormatter = DateFormatter()
+        dateFormatter.locale = .init(identifier: "ko_KR")
+        dateFormatter.dateFormat = "yyyy-MM-dd(EEEEE) a HH:mm"
+    
         cell.dateLabel.text = dateFormatter.string(from: date)
                 
         let iconName: String = weatherForecastInfo.weather.icon         
@@ -156,8 +128,10 @@ extension ViewController: UITableViewDelegate {
         tableView.deselectRow(at: indexPath, animated: true)
         
         let detailViewController: WeatherDetailViewController = WeatherDetailViewController()
-        detailViewController.weatherForecastInfo = weatherJSON?.weatherForecast[indexPath.row]
-        detailViewController.cityInfo = weatherJSON?.city
+        guard let weatherInfo = model.weatherJSON?.weatherForecast[indexPath.row] else { return }
+        
+        detailViewController.weatherForecastInfo = weatherInfo
+        detailViewController.cityInfo = model.weatherJSON?.city
         detailViewController.tempUnit = tempUnit
         navigationController?.show(detailViewController, sender: self)
     }
